major refactor, separated code between various packages.
