+++ app/router/route.go
@@ -1,0 +1,12 @@
+package router
+
+import "net/http"
+
+type Route struct {
+	Name        string
+	Pattern     string
+	Method      string
+	HandlerFunc http.HandlerFunc
+}
+
+type Routes []Route
