+++ app/router/router.go
@@ -1,0 +1,20 @@
+package router
+
+import (
+	"github.com/gorilla/mux"
+)
+
+func New() *mux.Router {
+
+	router := mux.NewRouter().StrictSlash(true)
+	for _, route := range routes {
+
+		router.
+			Methods(route.Method).
+			Path(route.Pattern).
+			Name(route.Name).
+			Handler(route.HandlerFunc)
+	}
+
+	return router
+}
