--- glide.lock
@@ -1,10 +1,0 @@
-hash: 370221a36f81046a7d2c6aa45e7114adc349804673f3f06e2fffef465aeedae0
-updated: 2018-11-10T08:46:36.709246Z
-imports:
-- name: github.com/gorilla/context
-  version: 51ce91d2eaddeca0ef29a71d766bb3634dadf729
-- name: github.com/gorilla/mux
-  version: e3702bed27f0d39777b0b37b664b6280e8ef8fbf
-- name: github.com/satori/go.uuid
-  version: f58768cc1a7a7e77a3bd49e98cdd21419399b6a3
-testImports: []
