--- glide.yaml
@@ -1,6 +1,0 @@
-package: .
-import:
-- package: github.com/gorilla/mux
-  version: ^1.6.2
-- package: github.com/satori/go.uuid
-  version: ^1.2.0
